Add required param check in update/delete user

+++ api/src/controller/users.controller.ts
@@ -22,6 +22,7 @@ export const createUser = async (req: Request, res: Response) => {
 
 export const updateUser = async (req: Request, res: Response) => {
     try {
+        if(!req.body.uuid) throw "User's UUID is mandatory"
         let users = await userService.updateUser(req.body.uuid as string, req.body)
         res.send(responseMessage.message(200, users))
     } catch (e) {
@@ -31,6 +32,7 @@ export const updateUser = async (req: Request, res: Response) => {
 
 export const deleteUser = async (req: Request, res: Response) => {
     try {
+        if(!req.body.uuid) throw "User's UUID is mandatory"
         let users = await userService.deleteUser(req.body.uuid as string)
         res.send(responseMessage.message(200, users))
     } catch (e) {
